Added overlay to components
