+++ src/Mamba/Buttons.js
@@ -153,6 +153,7 @@ export const OpenButton = styled.button`
   left: 0;
   right: 0;
   bottom: 0;
+  width: 100%;
   z-index: 1;
   color: inherit;
   transition: opacity 0.229s cubic-bezier(0.52, 0.16, 0.24, 1) 0.37s;
